Cleaned up script and renamed it.
